rename main to mainDynamic
