+++ mainDynamic.py
